fix(Input): ensure right touch controller is set to primary
Because the Oculus Input Action types use the OVRInput virtual
methods, it means any reference to the controller type is always
primary. Secondary is only ever used when referencing other controllers
such as the Xbox Gamepad or when using the OVRInput raw methods.

The prefab has been updated to the correct type now.