chore(Velocity): rename velocity estimator to more specific name
The Oculus Velocity Estimator isn't a generic estimator as it is
specifically tied to the OVRCameraRig Anchor points to get the
velocities for objects of that type.

Therefore, it's more appropriate to name it around what it's actually
doing and this lends itself better to determining what the component
actually does.